+++ clients/horizonclient/CHANGELOG.md
@@ -5,7 +5,10 @@ file.  This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
 
+### Add
+
 * Added transaction and operation result codes to the horizonclient.Error string for easy glancing at string only errors for underlying cause.
+* Added `IsHorizonAPITimeoutError` - returns whether the error indicates Horizon API has timed out.
 
 ## [v7.0.0](https://github.com/stellar/go/releases/tag/horizonclient-v7.0.0) - 2021-05-15
 
@@ -67,7 +70,7 @@ Added new client methods and effects supporting [Protocol 14](https://github.com
 
 ### Add
 
-- Add `IsNotFoundError`
+- Add `IsNotFoundError`.
 - Add `client.SubmitFeeBumpTransaction` and `client.SubmitFeeBumpTransactionWithOptions` for submitting [fee bump transactions](https://github.com/stellar/stellar-protocol/blob/master/core/cap-0015.md). Note that fee bump transactions will only be accepted by Stellar Core once Protocol 13 is enabled.
 
 ### Updates